- Hao's commit: Use int32 instead of int64 in mkl.--prepare of magma - However, this doesn't work with magma_lapack.h

+++ inc/matrix_define.h
@@ -4,7 +4,8 @@
 #ifdef USE_MKL
 #include "mkl.h"
 #define FORTRAN_NAME(x) x
-typedef  long long        BL_INT;
+//typedef  long long        BL_INT;
+typedef  int              BL_INT;
 typedef  float            BL_FLOAT;     
 typedef  double           BL_DOUBLE;    
 typedef  MKL_Complex8     BL_COMPLEX8;
